Pass iterator to conversion utilities
In using marctable in podbucket it was helpful to change the interface
of the conversion functions (to_parquet, etc) so that it accepts an
iterator of pymarc.Record objects rather than a file of MARC or MARCXML
data.

Warning: this may break your code if you are using marctable as a
library! I added a section to the README about how to use marctable as a
library. Hopefully this isn't too disruptive, but it's not at v1.0.0
yet, so everything about this should be considered unstable.

Along the way I split up the tests into separate files, and also started
using itertools.batched to do the batching of records which allowed me
to remove my own code for doing that.