Add Leader support
Adds a new column FLDR to the csv, jsonl and parquet outputs. This is an
intial implementation, since it doesn't include the different positions
split out into separate columns.

I think applications that need it can initially use `SUBSTR()` or
equivalent and see how far that gets? It may be that we need more
granularity, but I think it would be good to do that as part of
addressing #8 and #11 too?

Closes #15